backend/requirements.txt
Django==3.0.3
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
psycopg2==2.8.4
django-redis==4.11.0
hiredis==1.0.1
gunicorn==20.0.4
whitenoise==5.0.1
model-mommy==2.0.0
django_dramatiq==0.9.1
dramatiq[rabbitmq, watch]==1.8.1
requests==2.22.0
furl==2.1.0
